requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
yfinance>=0.1.87
requests>=2.28.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Technical Analysis
TA-Lib>=0.4.25

# Machine Learning
torch>=1.12.0
transformers>=4.21.0
joblib>=1.2.0
scipy>=1.9.0

# Google AI
google-generativeai>=0.3.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.8.0

# Caching
redis>=4.3.0

# Data processing
textblob>=0.17.1
nltk>=3.7

# Social Media APIs (optional)
tweepy>=4.14.0
praw>=7.6.0
asyncpraw>=7.7.0

# Utilities
python-dotenv>=0.20.0
pytz>=2022.1
psutil>=5.9.0
click>=8.1.0
warnings>=0.1
asyncio>=3.4.3

# Development dependencies
pytest>=7.1.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional dependencies for advanced features
plotly>=5.10.0
dash>=2.6.0
streamlit>=1.12.0